fix: CI 修复——watchdog 退避因子 u64 化、diverged 测试显式 clone -b main
- windows.rs：WATCHDOG_BACKOFF_FACTOR 从 u32 改为 u64，修复
  saturating_mul(as_secs()) 的类型不匹配（仅 Windows target 编译路径，
  本机 macOS 无法覆盖，由 CI Windows job 验证）
- branch_state diverged 测试：clone 显式 -b main。bare 远端 HEAD 未指向
  main 时，不同 git 版本的 clone fallback 不一致（CI 新版 git 落到
  unborn master，导致 push origin main 报 src refspec 不匹配），
  与既有测试的 clone 写法对齐

+++ src-tauri/src/desktop_widget/windows.rs
@@ -34,7 +34,8 @@ const WM_CREATE_DESKTOP_WORKER: u32 = 0x052C;
 const WATCHDOG_INTERVAL: Duration = Duration::from_secs(5);
 /// 连续失败时的退避上限，避免宿主异常时的重附风暴
 const WATCHDOG_BACKOFF_MAX: Duration = Duration::from_secs(80);
-const WATCHDOG_BACKOFF_FACTOR: u32 = 2;
+// u64：与 Duration::as_secs() 的返回类型一致（saturating_mul 参数要求同型）
+const WATCHDOG_BACKOFF_FACTOR: u64 = 2;
 /// 主窗口已退出时停止 watchdog 的哨兵错误（不可恢复，重试无意义）
 const WATCHDOG_STOP_SENTINEL: &str = "desktop_widget.watchdog.stop: main window closed";
